# EventManager/app/docs/swagger.yaml
basePath: /api/event-manager
definitions:
  hateoas.Link:
    properties:
      href:
        type: string
      method:
        type: string
      rel:
        type: string
      title:
        type: string
    type: object
...
  httpdto.HttpCreateTicket:
    properties:
      event_id:
        type: integer
      packet_id:
        type: integer
    type: object
...
  httpdto.HttpResponseTicket:
    properties:
      _links:
        additionalProperties:
          $ref: '#/definitions/hateoas.Link'
        type: object
      code:
        type: string
      event_id:
        type: integer
      packet_id:
        type: integer
    type: object
...
  httpdto.HttpUpdateTicket:
    properties:
      event_id:
        type: integer
      packet_id:
        type: integer
    type: object
...
host: localhost:12345
info:
  contact: {}
  description: Event management service for the POS system. Handles events, event
    packets, tickets, and inclusions.
  title: EventManager Service API
  version: "1.0"
paths:
...
  /tickets:
    post:
      consumes:
      - application/json
      description: Create a new ticket for an event and packet
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ticket details
        in: body
        name: ticket
        required: true
        schema:
          $ref: '#/definitions/httpdto.HttpCreateTicket'
      produces:
      - application/json
      responses:
        "201":
          description: Ticket created successfully
          schema:
            $ref: '#/definitions/httpdto.HttpResponseTicket'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - missing or invalid token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Event or packet not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Ticket already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new ticket
      tags:
      - tickets
  /tickets/{code}:
    delete:
      consumes:
      - application/json
      description: Delete a ticket by its code
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ticket code (UUID)
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Ticket deleted successfully
          schema:
            $ref: '#/definitions/httpdto.HttpResponseTicket'
        "400":
          description: Invalid ticket code
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - missing or invalid token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Ticket not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a ticket
      tags:
      - tickets
    get:
      consumes:
      - application/json
      description: Retrieve a specific ticket by its unique code
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ticket code (UUID)
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Ticket details
          schema:
            $ref: '#/definitions/httpdto.HttpResponseTicket'
        "400":
          description: Invalid ticket code
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - missing or invalid token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Ticket not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get ticket by code
      tags:
      - tickets
    patch:
      consumes:
      - application/json
      description: Partially update ticket details (PATCH - only provided fields are
        updated)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ticket code (UUID)
        in: path
        name: code
        required: true
        type: string
      - description: Fields to update
        in: body
        name: ticket
        required: true
        schema:
          $ref: '#/definitions/httpdto.HttpUpdateTicket'
      produces:
      - application/json
      responses:
        "200":
          description: Ticket updated successfully
          schema:
            $ref: '#/definitions/httpdto.HttpResponseTicket'
        "400":
          description: Invalid request body or ticket code
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - missing or invalid token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Ticket not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a ticket
      tags:
      - tickets
    put:
      consumes:
      - application/json
      description: Create a new ticket with a specified code or replace an existing
        one (idempotent)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ticket code (UUID)
        in: path
        name: code
        required: true
        type: string
      - description: Ticket details
        in: body
        name: ticket
        required: true
        schema:
          $ref: '#/definitions/httpdto.HttpCreateTicket'
      produces:
      - application/json
      responses:
        "200":
          description: Ticket created or updated successfully
          schema:
            $ref: '#/definitions/httpdto.HttpResponseTicket'
        "400":
          description: Invalid request body or ticket code
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - missing or invalid token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Event or packet not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create or replace a ticket with specific code
      tags:
      - tickets
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"